+++ src/software/simulated_tests/simulated_action_tests/move_action_test.cpp
@@ -5,7 +5,7 @@
 #include <utility>
 
 #include "software/geom/algorithms/contains.h"
-#include "software/simulated_tests/simulated_action_tests/simulated_action_test_fixture.h"
+#include "software/simulated_tests/simulated_action_test_fixture.h"
 #include "software/simulated_tests/validation/validation_function.h"
 #include "software/simulated_tests/validation_functions/robot_at_position_validation.h"
 #include "software/test_util/test_util.h"
